moved schemas to own schema folder, and added package-lock.json to .gitignore
