+++ src/server/api/question-game/game/schema/create.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+export const InputSchema = z.object({
+    title: z.string()
+});
+
+export const OutputSchema = z.object({
+    questionGameId: z.string()
+});
